Digital_resume/requirements.txt
requests
python-dotenv
gradio
openai
openai-agents
langchain_openai
langchain_chroma
numpy